[c++] Add `SOMABaseNDArray::shape`
