+++ libtiledbsoma/src/CMakeLists.txt
@@ -63,6 +63,7 @@ add_library(TILEDB_SOMA_OBJECTS OBJECT
   ${CMAKE_CURRENT_SOURCE_DIR}/soma/soma_measurement.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/soma/soma_context.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/soma/soma_dataframe.cc
+  ${CMAKE_CURRENT_SOURCE_DIR}/soma/soma_base_ndarray.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/soma/soma_dense_ndarray.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/soma/soma_sparse_ndarray.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/soma/array_buffers.cc
@@ -208,6 +209,7 @@ install(FILES
   ${CMAKE_CURRENT_SOURCE_DIR}/soma/soma_group.h
   ${CMAKE_CURRENT_SOURCE_DIR}/soma/soma_collection.h
   ${CMAKE_CURRENT_SOURCE_DIR}/soma/soma_dataframe.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/soma/soma_base_ndarray.h
   ${CMAKE_CURRENT_SOURCE_DIR}/soma/soma_dense_ndarray.h
   ${CMAKE_CURRENT_SOURCE_DIR}/soma/soma_sparse_ndarray.h
   ${CMAKE_CURRENT_SOURCE_DIR}/soma/soma_experiment.h
