+++ libtiledbsoma/src/soma/soma_base_ndarray.cc
@@ -1,0 +1,113 @@
+/**
+ * @file   soma_base_ndarray.cc
+ *
+ * @section LICENSE
+ *
+ * The MIT License
+ *
+ * @copyright Copyright (c) 2022-2023 TileDB, Inc.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ *
+ * @section DESCRIPTION
+ *
+ *   This file defines the SOMABaseNDArray class.
+ */
+
+#include "soma_base_ndarray.h"
+
+namespace tiledbsoma {
+using namespace tiledb;
+
+std::vector<int64_t> SOMABaseNDArray::shape() {
+    std::vector<int64_t> result;
+    auto dimensions = tiledb_schema()->domain().dimensions();
+
+    for (const auto& dim : dimensions) {
+        switch (dim.type()) {
+            case TILEDB_UINT8:
+                result.push_back(
+                    dim.domain<uint8_t>().second - dim.domain<uint8_t>().first +
+                    1);
+                break;
+            case TILEDB_INT8:
+                result.push_back(
+                    dim.domain<int8_t>().second - dim.domain<int8_t>().first +
+                    1);
+                break;
+            case TILEDB_UINT16:
+                result.push_back(
+                    dim.domain<uint16_t>().second -
+                    dim.domain<uint16_t>().first + 1);
+                break;
+            case TILEDB_INT16:
+                result.push_back(
+                    dim.domain<int16_t>().second - dim.domain<int16_t>().first +
+                    1);
+                break;
+            case TILEDB_UINT32:
+                result.push_back(
+                    dim.domain<uint32_t>().second -
+                    dim.domain<uint32_t>().first + 1);
+                break;
+            case TILEDB_INT32:
+                result.push_back(
+                    dim.domain<int32_t>().second - dim.domain<int32_t>().first +
+                    1);
+                break;
+            case TILEDB_UINT64:
+                result.push_back(
+                    dim.domain<uint64_t>().second -
+                    dim.domain<uint64_t>().first + 1);
+                break;
+            case TILEDB_INT64:
+            case TILEDB_DATETIME_YEAR:
+            case TILEDB_DATETIME_MONTH:
+            case TILEDB_DATETIME_WEEK:
+            case TILEDB_DATETIME_DAY:
+            case TILEDB_DATETIME_HR:
+            case TILEDB_DATETIME_MIN:
+            case TILEDB_DATETIME_SEC:
+            case TILEDB_DATETIME_MS:
+            case TILEDB_DATETIME_US:
+            case TILEDB_DATETIME_NS:
+            case TILEDB_DATETIME_PS:
+            case TILEDB_DATETIME_FS:
+            case TILEDB_DATETIME_AS:
+            case TILEDB_TIME_HR:
+            case TILEDB_TIME_MIN:
+            case TILEDB_TIME_SEC:
+            case TILEDB_TIME_MS:
+            case TILEDB_TIME_US:
+            case TILEDB_TIME_NS:
+            case TILEDB_TIME_PS:
+            case TILEDB_TIME_FS:
+            case TILEDB_TIME_AS:
+                result.push_back(
+                    dim.domain<int64_t>().second - dim.domain<int64_t>().first +
+                    1);
+                break;
+            default:
+                throw TileDBSOMAError("Dimension must be integer type.");
+        }
+    }
+
+    return result;
+}
+}  // namespace tiledbsoma
