+++ libtiledbsoma/src/soma/soma_base_ndarray.h
@@ -1,0 +1,94 @@
+/**
+ * @file   soma_base_ndarray.h
+ *
+ * @section LICENSE
+ *
+ * The MIT License
+ *
+ * @copyright Copyright (c) 2022-2024 TileDB, Inc.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ *
+ * @section DESCRIPTION
+ *
+ *   This declares the SOMABaseNDArray class.
+ */
+
+#ifndef SOMA_BASE_NDARRAY
+#define SOMA_BASE_NDARRAY
+
+#include "soma_array.h"
+
+namespace tiledbsoma {
+
+class SOMABaseNDArray : public SOMAArray {
+   public:
+    //===================================================================
+    //= public non-static
+    //===================================================================
+
+    /**
+     * @brief Construct a new SOMABaseNDArray object.
+     *
+     * @param mode read or write
+     * @param uri URI of the array
+     * @param name name of the array
+     * @param platform_config Config parameter dictionary
+     * @param column_names Columns to read
+     * @param batch_size Batch size
+     * @param result_order Result order
+     * @param timestamp Timestamp
+     */
+    SOMABaseNDArray(
+        OpenMode mode,
+        std::string_view uri,
+        std::shared_ptr<SOMAContext> ctx,
+        std::string_view name,
+        std::vector<std::string> column_names,
+        std::string_view batch_size,
+        ResultOrder result_order,
+        std::optional<TimestampRange> timestamp = std::nullopt)
+        : SOMAArray(
+              mode,
+              uri,
+              ctx,
+              name,
+              column_names,
+              batch_size,
+              result_order,
+              timestamp) {
+    }
+
+    SOMABaseNDArray(const SOMAArray& other)
+        : SOMAArray(other) {
+    }
+
+    using SOMAArray::open;
+
+    /**
+     * @brief Get the capacity of each dimension.
+     *
+     * @return A vector with length equal to the number of dimensions; each
+     * value in the vector is the capcity of each dimension.
+     */
+    std::vector<int64_t> shape();
+};
+}  // namespace tiledbsoma
+
+#endif  // SOMA_BASE_NDARRAY
